+++ CHANGES.md
@@ -3,8 +3,10 @@
 =====
 * Updated README.md to include MPI instructions.
 * splatt-check now outputs 1-indexed .map files.
+* splatt-check now sums duplicate nonzeros.
 * Memory allocations are now 64-byte aligned with `posix_memalign`. Memory
   allocation can be adjusted by altering `splatt_malloc`.
+* Fixed extern "C" typo in include/splatt.h
 
 
 1.1.0